move the decoding of a VaultFile name from datona-blockchain to datona-vault
The decoding and validating of a VaultFile filename was encapsulated in the wrong class - in the Contract class instead of the VaultKeeper class.  This is why the VaultFile class had to be held in a new types.js file when it clearly belongs in the datona-vault comonent.

Moved the VaultFile class from types to datona-vault and deleted the types file.
Moved the decode and validation of the vault file name from the Contract class to the VaultKeeper class.
Updated the docs to suit.