Change append permission stategy for files in directories
Previously, append permission for a file in a directory was directly inherited from the directory's permissions.  This change makes the append permission more intuitive and more flexible.  If a directory has write permissions then files within that directory can be unconditionally appended to.  If the directory does not have write permissions but has append permissions then new files can be appended to the directory but existing files cannot be written to or appended to.

This change modifies the VaultDataServer interface so vault servers that use this version will need to change.  The only change to the interface is adding a new method called 'createFile', which is the same as 'write' except it must reject if the file already exists.